@@ -1,5 +1,5 @@
 {
-  "homepage": "https://tickle17.github.io/DornetStore/",
+  "homepage": "https://github.com/Tickle17/dornetstore",
   "name": "dornet_shop",
   "version": "0.1.0",
   "private": true,
@@ -12,8 +12,8 @@
     "@mui/styles": "^5.12.0",
     "axios": "^1.4.0",
     "gh-pages": "^5.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-loadable": "^5.5.0",
     "react-scripts": "5.0.1",
     "react-slick": "^0.29.0",
