@@ -1,18 +1,18 @@
 import React, { lazy, Suspense } from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import Loader from "./loader"
 
 const App = lazy(() => import('./App'));
 
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
    <React.StrictMode>
       <React.StrictMode>
          <Suspense fallback={<Loader />}>
             <App />
          </Suspense>
       </React.StrictMode>,
-   </React.StrictMode>,
-document.getElementById('root')
+   </React.StrictMode>
 );
